Strip double quotes from highlighted words.
Double quotes seem to mean "do an exact match" instead of searching for
literal quotes in the text, so it makes sense to not highlight the quotes
themselves.

Leaving the quotes in the words to be highlighted also causes a problem if you
search for something like "high and low" -- it will highlight an instance of
the word high with <span class="highlight"> and then will also match the "high
in class="highlight" and wrap *that* with a highlight tag. You end up getting
something like:

    <span class="<span class="highlight">high</span>light">high</span> and low